Initial LibTransport with GetLiveTraffic
Attempting a bit of an architecture here where we have a Lib with a bunch of functions and method on them so that it can be used similar to a third party SDK. The idea is to make the Lib seem straight forward to use, and that the lib gives you data in types that will make life easier for consumers on the other side (i.e the UI app itself).

There is still a use case / command pattern going on here, but instead of super separating them out and unit testing each individual component, I'm instead doing more integration tests where we setup responses from the MockWebServer and test them out all the way through to the subscriber.

I decided to use Rx only because it's a tool I've been using quite a bit recently in a similar architecture, and I wanted to see how it would play out in this scenario and whether I could try and make it seem a bit nicer to use.

As for the entities, I started trying to make a bit of a 'smart' decoder for all the GeoJson, but the most important parts we really want are the properties from each feature if there there. The TransportNSW open data site has some docs for each endpoint and what the properties that are being returned are, so we can use that as our docs for knowing that certain properties are there. I'm taking a bit of a gamble assuming that all the properties will be there, even though they may not be, but I'll allow it for now.