Setup Koin and pass around a LibTransport
I _really_ didn't want to spend time writing custom factories for each view model that I'm going to be having, plus having to do some extra stuff in order to get args through. I had used Koin previously and it was quite nice, plus they have support for SavedStateHandle.

With this setup, I only need to list the viewmodel in our modules file with a bunch of get()'s, and then it's all done. I vastly prefer this to having to write a factory for each view model just so i can have it. I was hoping to avoid Dagger (which I have done) and any other DI tool on the UI, just to see how far I could get without them, but this is just too good to pass up